Postgres

+++ mcp-server-azure-resourcemanager-resources/src/main/java/org/agoncal/sample/mcp/azure/resourcemanager/resources/AzureResourceManagerResourcesMCPTools.java
@@ -16,30 +16,39 @@ public class AzureResourceManagerResourcesMCPTools {
 
     private static final Logger log = Logger.getLogger(AzureResourceManagerResourcesMCPTools.class);
 
-    @Tool(name = "creates_a_resource_group", description = "Creates a new resource group. If the resource group already exists, the operation fails")
-    public ToolResponse createResourceGroup(@ToolArg(name = "resource_group_name", description = "The name of the resource group to be created. A resource group in Azure is a container that holds related resources (storage account, database, message hubs...). The name of the resource group cannot have spaces and should start with the prefix 'rg-'.") String resourceGroupName, McpLog mcpLog) {
+    @Tool(name = "creates_a_resource_group", description = "Creates a new Resource Group in Azure. If the Resource Group already exists, the operation succeeds silently.")
+    public ToolResponse createResourceGroup(@ToolArg(name = "resource_group_name", description = "The name of the Resource Group to be created. A Resource Group in Azure is a container that holds related resources (storage account, database, message hubs...). The name of the Resource Group cannot have spaces and should start with the prefix 'rg-'.") String resourceGroupName, McpLog mcpLog) {
         log.info("Creating a resource group: " + resourceGroupName);
 
         AzureResourceManager azure = getAzureResourceManager();
 
-        ResourceGroup resourceGroup = azure.resourceGroups().define(resourceGroupName)
-            .withRegion(Region.US_EAST)
-            .create();
+        if (azure.resourceGroups().contain(resourceGroupName)) {
+            mcpLog.error("Not creating resource group " + resourceGroupName + " because it already exists");
+        } else {
+            ResourceGroup resourceGroup = azure.resourceGroups().define(resourceGroupName)
+                .withRegion(Region.US_EAST)
+                .create();
 
-        mcpLog.info("Resource Group " + resourceGroup.name() + " has been created");
+            mcpLog.info("Resource Group " + resourceGroup.name() + " has been created");
+        }
         return ToolResponse.success();
     }
 
-    @Tool(name = "deletes_a_resource_group", description = "Deletes an existing resource group. If the resource group does not exists, the operation fails")
-    public ToolResponse deleteResourceGroup(@ToolArg(name = "resource_group_name", description = "The name of the resource group to be deleted. .") String resourceGroupName, McpLog mcpLog) {
+    @Tool(name = "deletes_a_resource_group", description = "Deletes an existing Resource Group from Azure. If the Resource Group does not exists, the operation fails")
+    public ToolResponse deleteResourceGroup(@ToolArg(name = "resource_group_name", description = "The name of the Resource Group to be deleted. .") String resourceGroupName, McpLog mcpLog) {
         log.info("Deleting a resource group: " + resourceGroupName);
 
         AzureResourceManager azure = getAzureResourceManager();
 
-        azure.resourceGroups().deleteByName(resourceGroupName);
+        if (azure.resourceGroups().contain(resourceGroupName)) {
+            mcpLog.error("Not deleting resource group " + resourceGroupName + " because it does not exist");
+            return ToolResponse.error("Not deleting resource group " + resourceGroupName + " because it does not exist");
+        } else {
+            azure.resourceGroups().deleteByName(resourceGroupName);
 
-        mcpLog.info("Resource Group " + resourceGroupName + " has been deleted");
-        return ToolResponse.success();
+            mcpLog.info("Resource Group " + resourceGroupName + " has been deleted");
+            return ToolResponse.success();
+        }
     }
 
     private static AzureResourceManager getAzureResourceManager() {
